refactor(lib): Create lib/domain/ for business logic modules
- Create lib/domain/ directory
- Move lib/content → lib/domain/content (31 imports updated)
- Move lib/editor → lib/domain/editor (2 imports updated)
- Move lib/search → lib/domain/search (4 imports updated)
- Move lib/admin → lib/domain/admin (17 imports updated)
- Update all import paths across codebase (54 total imports)
- Rationale: Group business logic modules for better discoverability
