+++ app/api/admin/audit-logs/export/route.ts
@@ -9,8 +9,8 @@
 import { NextRequest } from "next/server";
 import { prisma } from "@/lib/database/client";
 import { requireRole } from "@/lib/auth/middleware";
-import { logAuditAction, generateCSV, handleApiError } from "@/lib/admin/audit";
-import { AUDIT_ACTIONS } from "@/lib/admin/api-types";
+import { logAuditAction, generateCSV, handleApiError } from "@/lib/domain/admin/audit";
+import { AUDIT_ACTIONS } from "@/lib/domain/admin/api-types";
 
 // ============================================================
 // GET /api/admin/audit-logs/export - Export CSV
