+++ app/api/admin/content/[id]/route.ts
@@ -13,8 +13,8 @@ import {
   logAuditAction,
   handleApiError,
   deriveContentType,
-} from "@/lib/admin/audit";
-import { AUDIT_ACTIONS, type AdminContentDetail } from "@/lib/admin/api-types";
+} from "@/lib/domain/admin/audit";
+import { AUDIT_ACTIONS, type AdminContentDetail } from "@/lib/domain/admin/api-types";
 
 // ============================================================
 // GET /api/admin/content/[id] - Content Detail
