+++ app/api/content/content/tree/route.ts
@@ -9,7 +9,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/database/client";
 import { requireAuth } from "@/lib/auth/middleware";
-import { deriveContentType } from "@/lib/content";
+import { deriveContentType } from "@/lib/domain/content";
 
 // ============================================================
 // GET /api/content/content/tree - Get Content Tree
