+++ app/api/content/storage/[id]/route.ts
@@ -9,7 +9,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/database/client";
 import { requireAuth } from "@/lib/auth/middleware";
-import type { UpdateStorageConfigRequest } from "@/lib/content/api-types";
+import type { UpdateStorageConfigRequest } from "@/lib/domain/content/api-types";
 
 type Params = Promise<{ id: string }>;
