+++ app/api/content/tags/extract/route.ts
@@ -18,7 +18,7 @@
 
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/database/client";
-import { extractTags } from "@/lib/content/tag-extractor";
+import { extractTags } from "@/lib/domain/content/tag-extractor";
 import type { JSONContent } from "@tiptap/core";
 
 interface ExtractTagsRequest {
