+++ components/content/FileTree.tsx
@@ -18,7 +18,7 @@ import { useRef, useEffect, useMemo } from "react";
 import { Tree, type NodeApi } from "react-arborist";
 import { FileNode } from "./FileNode";
 import { useTreeStateStore } from "@/state/tree-state-store";
-import type { TreeNode } from "@/lib/content/types";
+import type { TreeNode } from "@/lib/domain/content/types";
 
 interface FileTreeProps {
   data: TreeNode[];
