+++ components/content/FileTreeWithDropZone.tsx
@@ -6,7 +6,7 @@ import { NativeTypes } from "react-dnd-html5-backend";
 import { validateFileBatch } from "@/lib/media/file-validation";
 import { toast } from "sonner";
 import { FileTree } from "./FileTree";
-import type { TreeNode } from "@/lib/content/types";
+import type { TreeNode } from "@/lib/domain/content/types";
 
 interface FileTreeWithDropZoneProps {
   data: TreeNode[];
