+++ components/content/OutlinePanel.tsx
@@ -16,7 +16,7 @@
 
 import { useState, useEffect } from "react";
 import { useContentStore } from "@/state/content-store";
-import type { OutlineHeading } from "@/lib/content/outline-extractor";
+import type { OutlineHeading } from "@/lib/domain/content/outline-extractor";
 
 interface OutlinePanelProps {
   /** Current outline headings from the editor */
