+++ lib/domain/content/index.ts
