+++ lib/domain/content/metadata-validation.ts
