+++ lib/domain/content/outline-extractor.ts
