+++ lib/domain/content/search-text.ts
