+++ lib/domain/content/slug.ts
