+++ lib/domain/content/types.ts
