+++ lib/domain/editor/bullet-list-backspace.ts
