+++ lib/domain/editor/extensions-server.ts
